+++ PSYDataAdditions/PSYDataScanner.m
@@ -69,7 +69,7 @@
 }
 #endif
 
-- (void)setScanLocation:(NSUInteger)value
+- (void)setScanLocation:(unsigned long long)value
 {
     if(value > [_scannedData length])
         [NSException raise:NSRangeException format:@"*** -[PSYDataScanner setScanLocation:]: Range or index out of bounds"];
@@ -105,7 +105,7 @@
 
 - (BOOL)scanInt8:(uint8_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL) [_scannedData getBytes:value range:NSMakeRange(_scanLocation, length)];
@@ -116,7 +116,7 @@
 
 - (BOOL)scanLittleEndianInt16:(uint16_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -131,7 +131,7 @@
 
 - (BOOL)scanLittleEndianInt32:(uint32_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -146,7 +146,7 @@
 
 - (BOOL)scanLittleEndianInt64:(uint64_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -161,7 +161,7 @@
 
 - (BOOL)scanBigEndianInt16:(uint16_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -176,7 +176,7 @@
 
 - (BOOL)scanBigEndianInt32:(uint32_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -191,7 +191,7 @@
 
 - (BOOL)scanBigEndianInt64:(uint64_t *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -241,7 +241,7 @@
 
 - (BOOL)scanLittleEndianVarint32:(uint32_t *)value
 {
-    NSUInteger loc = [self scanLocation];
+    unsigned long long loc = [self scanLocation];
     
     uint32_t result = 0;
     
@@ -268,7 +268,7 @@
 
 - (BOOL)scanLittleEndianVarint64:(uint64_t *)value
 {
-    NSUInteger loc = [self scanLocation];
+    unsigned long long loc = [self scanLocation];
     
     uint64_t result = 0;
 
@@ -295,7 +295,7 @@
 
 - (BOOL)scanBigEndianVarint32:(uint32_t *)value
 {
-    NSUInteger loc = [self scanLocation];
+    unsigned long long loc = [self scanLocation];
     
     int8_t shift;
     uint32_t result = 0;
@@ -341,7 +341,7 @@
 
 - (BOOL)scanBigEndianVarint64:(uint64_t *)value
 {
-    NSUInteger loc = [self scanLocation];
+    unsigned long long loc = [self scanLocation];
     
     int8_t shift;
     uint64_t result = 0;
@@ -482,7 +482,7 @@
 
 - (BOOL)scanFloat:(float *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL) [_scannedData getBytes:value range:NSMakeRange(_scanLocation, length)];
@@ -493,7 +493,7 @@
 
 - (BOOL)scanDouble:(double *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL) [_scannedData getBytes:value range:NSMakeRange(_scanLocation, length)];
@@ -504,7 +504,7 @@
 
 - (BOOL)scanSwappedFloat:(float *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -520,7 +520,7 @@
 
 - (BOOL)scanSwappedDouble:(double *)value
 {
-    NSUInteger length = sizeof(*value);
+    unsigned long long length = sizeof(*value);
     if(_scanLocation + length > _dataLength) return NO;
     
     if(value != NULL)
@@ -534,7 +534,7 @@
     return YES;
 }
 
-- (BOOL)scanData:(NSData **)data ofLength:(NSUInteger)length
+- (BOOL)scanData:(NSData **)data ofLength:(unsigned long long)length
 {
     if(_scanLocation + length > _dataLength) return NO;
     
@@ -547,7 +547,7 @@
 
 - (BOOL)scanData:(NSData *)data intoData:(NSData **)dataValue
 {
-    NSUInteger length = [data length];
+    unsigned long long length = [data length];
     if(_scanLocation + length > _dataLength) return NO;
     
     if(length > 0)
@@ -568,7 +568,7 @@
 
 - (BOOL)scanUpToData:(NSData *)stopData intoData:(NSData **)dataValue
 {
-    NSUInteger length = [stopData length];
+    unsigned long long length = [stopData length];
     
     NSRange scannedRange = NSMakeRange(_scanLocation, 0);
     
@@ -581,10 +581,10 @@
         
         BOOL hasFoundData = NO;
         
-        for(NSUInteger scannedLoc = _scanLocation; scannedLoc + length <= _dataLength; scannedLoc++)
+        for(unsigned long long scannedLoc = _scanLocation; scannedLoc + length <= _dataLength; scannedLoc++)
         {
             hasFoundData = YES;
-            for(NSUInteger stopLoc = 0; stopLoc < length; stopLoc++)
+            for(unsigned long long stopLoc = 0; stopLoc < length; stopLoc++)
                 if(scannedBuffer[scannedLoc + stopLoc] != stopBuffer[stopLoc])
                 {
                     hasFoundData = NO;
@@ -612,7 +612,7 @@
     return YES;
 }
 
-- (BOOL)scanString:(NSString **)value ofLength:(NSUInteger)length usingEncoding:(NSStringEncoding)encoding
+- (BOOL)scanString:(NSString **)value ofLength:(unsigned long long)length usingEncoding:(NSStringEncoding)encoding
 {
     if(_scanLocation + length > _dataLength) return NO;
